Refactored sabre.js to support multiple execution flows. Introduced --apiVersion CLI option.
