+++ lib/client.js
@@ -99,7 +99,15 @@ const getMythXReport = async (apiUrl, ethAddress, password, data, initialDelay, 
     };
 };
 
+const getMythXApiVersion = async (apiUrl, ethAddress, password) => {
+    const client = new mythx.Client(ethAddress, password, undefined, apiUrl);
+    const data = await client.getVersion();
+
+    return data;
+};
+
 module.exports = {
+    getMythXApiVersion,
     getMythXReport,
     getRequestData
 };
