+++ lib/compiler.js
@@ -174,7 +174,7 @@ const getFunctionHashes = (contracts) => {
     }
 
     return hashes;
-}
+};
 
 /*
  * Compile contracts using solc snapshot
@@ -231,12 +231,6 @@ const getCompiledContracts = (input, solcSnapshot, solidityFileName, compileCont
     if (!contract.evm.bytecode.object) {
         throw new Error('✖ Compiling the Solidity code did not return any bytecode. Note that abstract contracts cannot be analyzed.');
     }
-
-    /* Bytecode should match the regex /^(0x)?([0-9a-fA-F]{2})+$/ */
-
-    // if (!/^(0x)?([0-9a-fA-F]{2})+$/.test(contract.evm.bytecode.object)) {
-    //     throw new Error('✖ Generated bytecode fails to match the required pattern: /^(0x)?([0-9a-fA-F]{2})+$/');
-    // }
 
     const functionHashes = getFunctionHashes(compiled.contracts);
