+++ lib/formatters/text.js
@@ -12,12 +12,12 @@ const textFormatter = {};
 textFormatter.strToInt = str => parseInt(str, 10);
 
 textFormatter.guessAccountRoleByAddress = (address) => {
-    switch (address.toLowerCase().substr(0, 10)) {
-        case '0xaffeaffe':
-            return roles.creator;
+    const prefix = address.toLowerCase().substr(0, 10);
 
-        case '0xdeadbeef':
-            return roles.attacker;
+    if (prefix === '0xaffeaffe') {
+        return roles.creator;
+    } else if (prefix === '0xdeadbeef') {
+        return roles.attacker;
     }
 
     return roles.other;
@@ -35,7 +35,7 @@ textFormatter.stringifyValue = (value) => {
     }
 
     return JSON.stringify(value);
-}
+};
 
 textFormatter.formatTestCaseSteps = (steps, fnHashes) => {
     const output = [];
@@ -94,14 +94,14 @@ textFormatter.getCodeSample = (source, src) => {
     const [start, length] = src.split(':').map(textFormatter.strToInt);
 
     return source.substr(start, length);
-}
+};
 
 textFormatter.formatLocation = message => {
     const start = message.line + ':' + message.column;
     const finish = message.endLine + ':' + message.endCol;
 
     return 'from ' + start + ' to ' + finish;
-}
+};
 
 textFormatter.formatMessage = (message, filePath, sourceCode, fnHashes) => {
     const { mythxIssue, mythxTextLocations } = message;
