+++ lib/report.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const eslintHelpers = require('./eslint');
-const eslintCliEngine = require("eslint").CLIEngine;
+const eslintCliEngine = require('eslint').CLIEngine;
 const SourceMappingDecoder = require('remix-lib/src/sourceMappingDecoder');
 
 const mythx2Severity = {
